rename arxivsearch and fix typing
