--- backend/arxiv/embeddings/providers/__init__.py
@@ -1,4 +1,0 @@
-from arxiv.embeddings.providers.cohere import CohereProvider
-
-
-__all__ = ["CohereProvider"]
