+++ backend/arxivsearch/data/redis-logo.jpg
