+++ backend/arxivsearch/data/redis-logo.png
