+++ backend/arxivsearch/embeddings/providers/__init__.py
@@ -1,0 +1,4 @@
+from arxivsearch.embeddings.providers.cohere import CohereProvider
+
+
+__all__ = ["CohereProvider"]
