+++ backend/arxivsearch/schema/__init__.py
@@ -1,8 +1,8 @@
-from arxiv.schema.request import (
+from arxivsearch.schema.request import (
     PaperSimilarityRequest,
     UserTextSimilarityRequest
 )
 
-from arxiv.schema.provider import Provider
+from arxivsearch.schema.provider import Provider
 
 __all__ = ["PaperSimilarityRequest", "UserTextSimilarityRequest", "Provider"]
