+++ backend/arxivsearch/schema/huggingface.yaml
