+++ backend/arxivsearch/schema/openai.yaml
