+++ backend/setup.cfg
@@ -1,5 +1,5 @@
 [metadata]
-name = arXiv_vecsim
+name = arxivsearch
 version = 0.0.1
 description = Workshop demo of Redis vector similarity over arXiv papers
 long_description = file: README.md
